migrate from activity to fragment
