+++ app/src/main/java/com/example/twitterminiapp/MainActivity.kt
@@ -2,10 +2,7 @@ package com.example.twitterminiapp
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Toast
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.twitterminiapp.data.util.Resource
 import com.example.twitterminiapp.databinding.ActivityMainBinding
 import com.example.twitterminiapp.presentation.adapter.TwitterAdapter
 import com.example.twitterminiapp.presentation.viewmodel.TwitterViewModel
@@ -24,46 +21,12 @@ class MainActivity : AppCompatActivity() {
     @Inject
     lateinit var viewModelFactory: TwitterViewModelFactory
 
-    private lateinit var viewModel: TwitterViewModel
+    lateinit var viewModel: TwitterViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
         viewModel = ViewModelProvider(this, viewModelFactory).get(TwitterViewModel::class.java)
-        initRecyclerView()
-        viewTimeline()
-    }
-
-    private fun viewTimeline() {
-        viewModel.getTimeline()
-        viewModel.tweets.observe(this, { resource ->
-            when (resource) {
-                is Resource.Success -> {
-                    resource.data?.let {
-                        twitterAdapter.differ.submitList(it)
-                    }
-                }
-                is Resource.Error -> {
-                    resource.message?.let {
-                        Toast.makeText(
-                            this,
-                            "An error occured: $it",
-                            Toast.LENGTH_LONG
-                        ).show()
-                    }
-                }
-                is Resource.Loading -> {
-                    // TODO Progress barを表示
-                }
-            }
-        })
-    }
-
-    private fun initRecyclerView() {
-        binding.homeTimelineRecyclerView.apply {
-            adapter = twitterAdapter
-            layoutManager = LinearLayoutManager(this@MainActivity)
-        }
     }
 }
